[TASK] Make aliased interfaces use class_alias
If interfaces (and only interfaces) are not aliased with class_alias,
an error is presented that the alias is not compatible with the
original - which is technically true because the loaded class name
is in fact replaced with the original name. Using class_alias this
way allows an alias of an interface to be used in method signatures.